Sanitize more of the embed iframe from youtube
Currently, when you try to scrape a youtube video with this via wallabag, the presence of url encoded quotes in the value of the iframe's `src` property is corrupting the rendered link.  An extra `https://www.youtube.com%20` is being inserted in said property at some point in the process.  This can be reproduced by having wallabag try to archive any youtube video (such as the one in this site config).

This change will make it so that the url encoded quotes are replaced with ASCII quotes, which is fixing the problem on my end.  I'm not sure if the behavior I'm seeing implies another bug somewhere in this process, or if I should be looking at another project, but as the intent of the site config appears to be to parse the iframe as if it were HTML, this seems like a needed fix either way.